move `/files/` redirect to base handlers
and reuse it in /nbconvert/html/…

closes #6137
