+++ IPython/html/nbconvert/handlers.py
@@ -9,7 +9,10 @@ import zipfile
 
 from tornado import web
 
-from ..base.handlers import IPythonHandler, notebook_path_regex
+from ..base.handlers import (
+    IPythonHandler, FilesRedirectHandler,
+    notebook_path_regex, path_regex,
+)
 from IPython.nbformat.current import to_notebook_json
 
 from IPython.utils.py3compat import cast_bytes
@@ -128,6 +131,7 @@ class NbconvertPostHandler(IPythonHandler):
 
         self.finish(output)
 
+
 #-----------------------------------------------------------------------------
 # URL to handler mappings
 #-----------------------------------------------------------------------------
@@ -139,4 +143,5 @@ default_handlers = [
     (r"/nbconvert/%s%s" % (_format_regex, notebook_path_regex),
          NbconvertFileHandler),
     (r"/nbconvert/%s" % _format_regex, NbconvertPostHandler),
+    (r"/nbconvert/html%s" % path_regex, FilesRedirectHandler),
 ]
